Restructure a lot of stuff
Hopefully won't regret doing this. Also, TODO: I'm not sure I like the
name src/components/core, would prefer something a bit better maybe?
