+++ src/components/core/plugin-container.tsx
