+++ src/components/experimental/change-project.ts
@@ -1,13 +1,13 @@
-import { getDirFiles, exists, pathRelativeToHome } from '../support/utils'
-import { createVim, renameCurrentToCwd } from '../core/instance-manager'
-import { RowNormal, RowImportant } from '../components/row-container'
-import { h, app, vimBlur, vimFocus } from '../ui/uikit'
-import { Plugin } from '../components/plugin-container'
-import Input from '../components/text-input'
+import { getDirFiles, exists, pathRelativeToHome } from '../../support/utils'
+import { createVim, renameCurrentToCwd } from '../../core/instance-manager'
+import { RowNormal, RowImportant } from '../core/row-container'
+import { h, app, vimBlur, vimFocus } from '../../ui/uikit'
+import { Plugin } from '../core/plugin-container'
+import Input from '../core/text-input'
 import { join, sep, basename } from 'path'
 import { filter } from 'fuzzaldrin-plus'
 import * as Icon from 'hyperapp-feather'
-import api from '../core/instance-api'
+import api from '../../core/instance-api'
 import { homedir } from 'os'
 
 const $HOME = homedir()
