+++ src/components/experimental/vim-create.ts
@@ -1,9 +1,9 @@
-import { Plugin } from '../components/plugin-container'
-import { app, vimBlur, vimFocus } from '../ui/uikit'
-import { createVim } from '../core/instance-manager'
-import Input from '../components/text-input'
+import { Plugin } from '../core/plugin-container'
+import { app, vimBlur, vimFocus } from '../../ui/uikit'
+import { createVim } from '../../core/instance-manager'
+import Input from '../core/text-input'
 import * as Icon from 'hyperapp-feather'
-import api from '../core/instance-api'
+import api from '../../core/instance-api'
 
 const state = {
   value: '',
