+++ src/components/legacy/divination.ts
