+++ src/components/legacy/inventory-search.ts
