+++ src/components/legacy/inventory.ts
