+++ src/components/legacy/problem-info.ts
