+++ src/components/legacy/spell-check.ts
