+++ src/components/lsp/hint.ts
@@ -1,9 +1,9 @@
-import { docStyle, resetMarkdownHTMLStyle } from '../ui/styles'
-import * as windows from '../windows/window-manager'
-import Overlay from '../components/overlay'
-import { h, app } from '../ui/uikit'
-import { cvar } from '../ui/css'
-import api from '../core/instance-api'
+import { docStyle, resetMarkdownHTMLStyle } from '../../ui/styles'
+import * as windows from '../../windows/window-manager'
+import Overlay from '../core/overlay'
+import { h, app } from '../../ui/uikit'
+import { cvar } from '../../ui/css'
+import api from '../../core/instance-api'
 import { parse as stringToMarkdown } from 'marked'
 
 interface ShowParams {
