+++ src/components/lsp/hover.ts
@@ -1,14 +1,14 @@
-import * as windows from '../windows/window-manager'
-import { ColorData } from '../services/colorizer'
-import { sub } from '../messaging/dispatch'
-import { debounce } from '../support/utils'
-import Overlay from '../components/overlay'
-import { docStyle } from '../ui/styles'
-import { cursor } from '../core/cursor'
-import { h, app } from '../ui/uikit'
+import * as windows from '../../windows/window-manager'
+import { ColorData } from '../../services/colorizer'
+import { sub } from '../../messaging/dispatch'
+import { debounce } from '../../support/utils'
+import Overlay from '../core/overlay'
+import { docStyle } from '../../ui/styles'
+import { cursor } from '../../core/cursor'
+import { h, app } from '../../ui/uikit'
 import { parse as stringToMarkdown, setOptions } from 'marked'
-import api from '../core/instance-api'
-import { cell, size as workspaceSize } from '../core/workspace'
+import api from '../../core/instance-api'
+import { cell, size as workspaceSize } from '../../core/workspace'
 
 setOptions({
   highlight: (code, lang, _) => {
