+++ src/components/nvim/statusline.ts
@@ -1,12 +1,12 @@
-import { onSwitchVim, getCurrentName } from '../core/instance-manager'
-import { getColorByName } from '../render/highlight-attributes'
-import { sub, processAnyBuffered } from '../messaging/dispatch'
-import { darken, brighten, cvar } from '../ui/css'
-import { ExtContainer } from '../neovim/protocol'
+import { onSwitchVim, getCurrentName } from '../../core/instance-manager'
+import { getColorByName } from '../../render/highlight-attributes'
+import { sub, processAnyBuffered } from '../../messaging/dispatch'
+import { darken, brighten, cvar } from '../../ui/css'
+import { ExtContainer } from '../../neovim/protocol'
 import * as Icon from 'hyperapp-feather'
-import api from '../core/instance-api'
-import { colors } from '../ui/styles'
-import { h, app } from '../ui/uikit'
+import api from '../../core/instance-api'
+import { colors } from '../../ui/styles'
+import { h, app } from '../../ui/uikit'
 import { basename } from 'path'
 import { homedir } from 'os'
