+++ src/render/events.ts
@@ -1,7 +1,7 @@
 import { CursorShape, setCursorColor, setCursorShape } from '../core/cursor'
 import { forceRegenerateFontAtlas } from '../render/font-texture-atlas'
-import { showMessageHistory } from '../components/message-history'
-import messages, { MessageKind } from '../components/messages'
+import { showMessageHistory } from '../components/nvim/message-history'
+import messages, { MessageKind } from '../components/nvim/messages'
 import { getColorById } from '../render/highlight-attributes'
 import { normalizeVimMode } from '../support/neovim-utils'
 import * as windows from '../windows/window-manager'
